FE-54: Retry-Dialog bei gescheitertem Request. Toast-Msgs, falls Netzwerkstatus geändert wird.
Patchset 3: Dem httpRequest wird nun mitgeteilt, welche Trackid und
welche Locationids er gerade versendet. Sobald der HttpRequest
failet, fügt er diese Daten in die Datenbank und startet einen
Intent. Dieser Intent öffnet die MainActivity. Diese zeigt dann
einen Dialog, in dem der Nutzer darauf hingewiesen wird, dass
Requests gefailet sind. Sobald der Nutzer auswählt, dass die Verbindung
neu versucht werden soll, werden die Datensätze aus der DB gelesen.

Patchset 4: Test gefixt.

+++ app/src/androidTest/java/de/h3adless/gpstracker/DatabaseTest.java
@@ -4,6 +4,7 @@ import android.database.sqlite.SQLiteDatabase;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -53,7 +54,7 @@ public class DatabaseTest {
 
 	private TrackPoint trackPoint;
 	private MainActivity mainActivity;
-	private int locationId;
+	private long locationId;
 
 	private boolean trackPointsAreEqual(TrackPoint one, TrackPoint two) {
 		return one.lat == two.lat &&
